Improving cache implementation
